requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyarrow>=14.0
PyYAML>=6.0
transformers>=4.36
